feat(state)!: no v0 migration — refuse pre-v1 state with instructions
Luke's call, and the right one: v0 shipped to nobody, which was the entire point of the v1 breaking window — owning a migration for an install base of zero is dead weight. The ~90-line v0->v1 migration is deleted along with its test and the rig purge helper's kv-blob fallback.

What replaces it is a courtesy, not a migration: opening a pre-v1 state.db now throws one actionable sentence (stop, delete the data volume, pair again) instead of raw no-such-column SQL errors. Unknown FUTURE schema versions are likewise refused rather than guessed at. Both refusals are unit-tested.
