--- node_modules
@@ -1,1 +1,0 @@
-/Users/luke/Documents/immich-shared-albums/node_modules
